requirements/base.txt
apistar
apistar-cors
apistar_jwt
fire
python-dotenv
gunicorn
watchtower
sqlalchemy
alembic
